Refactor js files into more fitting names
