+++ typing-practice-react/src/createStringForStats.js
