+++ typing-practice-react/src/timeHandler.js
@@ -1,11 +1,11 @@
-import {displayTodayStats,  displayTotalStats} from "./displayStats.js";
+import {createStringForTodayStats,  createStringForTotalStats} from "./createStringForStats.js";
 
 export function incrementSeconds(seconds, statsTextForSeconds, todayStatsTextContainer, totalStatsTextContainer) {
   seconds += 1;
   localStorage.todayTotalSeconds = parseInt(localStorage.todayTotalSeconds) + 1;
   localStorage.totalTotalSeconds = parseInt(localStorage.totalTotalSeconds) + 1;
-  todayStatsTextContainer.innerHTML = displayTodayStats();
-  totalStatsTextContainer.innerHTML = displayTotalStats();
+  todayStatsTextContainer.innerHTML = createStringForTodayStats();
+  totalStatsTextContainer.innerHTML = createStringForTotalStats();
   statsTextForSeconds.innerHTML = 'Seconds: ' + seconds + 's';
   resetAtMidnight();
